requirements.txt
canonicalwebteam.flask-base==1.1.0
canonicalwebteam.templatefinder==1.0.0
canonicalwebteam.image-template==1.3.1
